+++ src/index.js
@@ -8,8 +8,6 @@ let packagesList = config.packages
 let components = {} // 组件
 let methods = {} // 方法 $挂载
 
-console.log(packagesList)
-
 packagesList.map(item => {
   const pkg = packages[item.name]
 
@@ -40,7 +38,7 @@ let install = function (Vue, options = {}) {
   /**
    * 安装组件，便于全局直接引用在template中
    */
-  console.log(components)
+  // console.log(components)
   for (const componentKey in components) {
     components[componentKey] && components[componentKey].name && Vue.component(components[componentKey].name, components[componentKey])
   }