+++ src/packages/ActionSheet/ActionSheet.vue
@@ -1,31 +1,31 @@
 <template>
-  <div class="u-actionsheet">
+  <div class="lina-actionsheet">
     <transition name="f-fade">
-      <div class="u-actionsheet-mask" @click="clickActionSheetMask" v-show="isShow"></div>
+      <div class="lina-actionsheet-mask" @click="clickActionSheetMask" v-show="isShow"></div>
     </transition>
     <transition name="f-slide-up">
-      <div class="u-actionsheet-panel" v-show="isShow">
-        <div class="u-actionsheet-custom" v-if="$slots.custom">
+      <div class="lina-actionsheet-panel" v-show="isShow">
+        <div class="lina-actionsheet-custom" v-if="$slots.custom">
           <slot name="custom" v-html="custom"></slot>
         </div>
-        <dl class="u-actionsheet-modal" v-if="$slots.title || $slots.subTitle ">
-          <dt class="u-actionsheet-title">
+        <dl class="lina-actionsheet-modal" v-if="$slots.title || $slots.subTitle ">
+          <dt class="lina-actionsheet-title">
             <slot name="title" v-html="title"></slot>
           </dt>
-          <dd class="u-actionsheet-sub-title">
+          <dd class="lina-actionsheet-sub-title">
             <slot name="sub-title" v-html="subTitle"></slot>
           </dd>
         </dl>
-        <ul class="u-actionsheet-menu">
+        <ul class="lina-actionsheet-menu">
           <li
-            class="u-actionsheet-item"
-            :class="{'u-actionsheet-item-active': isHighlight(item)}"
+            class="lina-actionsheet-item"
+            :class="{'lina-actionsheet-item-active': isHighlight(item)}"
             v-for="(item, index) of menuItems"
             :key="index"
             @click="chooseItem(item)"
           >{{item[optionTag]}}</li>
         </ul>
-        <div class="u-actionsheet-cancel" v-if="cancelText" @click="closeActionSheet">{{cancelText}}</div>
+        <div class="lina-actionsheet-cancel" v-if="cancelText" @click="closeActionSheet">{{cancelText}}</div>
       </div>
     </transition>
   </div>
@@ -33,7 +33,7 @@
 
 <script>
 export default {
-  name: 'ActionSheet',
+  name: 'lina-action-sheet',
   props: {
     isShow: {
       type: Boolean,
@@ -103,14 +103,14 @@ export default {
 <!-- Add "scoped" attribute to limit CSS to this component only -->
 <style lang="scss" scope>
 //弹窗
-.u-actionsheet {
+.lina-actionsheet {
   text-align: center;
 
-  .u-actionsheet-mask {
+  .lina-actionsheet-mask {
     @include fix-fullscreen();
   }
 
-  .u-actionsheet-panel {
+  .lina-actionsheet-panel {
     position: fixed;
     left: 0;
     right: 0;
@@ -119,33 +119,33 @@ export default {
     z-index: 30;
     background-color: #f6f6f6;
   }
-  .u-actionsheet-modal {
+  .lina-actionsheet-modal {
     padding: 20px;
     margin: 0;
     text-align: center;
     background-color: #fff;
     border-bottom: 1px solid #F6F6F6;
-    .u-actionsheet-title,
-    .u-actionsheet-sub-title {
+    .lina-actionsheet-title,
+    .lina-actionsheet-sub-title {
       padding: 10px 0;
     }
-    .u-actionsheet-title {
+    .lina-actionsheet-title {
       font-size: 28px;
       color: #2D2D2D;
     }
-    .u-actionsheet-sub-title {
+    .lina-actionsheet-sub-title {
       font-size: 24px;
       color: #2D2D2D;
       margin-inline-start: 0px;
     }
   }
-  .u-actionsheet-menu {
+  .lina-actionsheet-menu {
     list-style: none;
     padding: 0;
     margin: 0;
   }
-  .u-actionsheet-cancel,
-  .u-actionsheet-item {
+  .lina-actionsheet-cancel,
+  .lina-actionsheet-item {
     height: 96px;
     padding: 0 20px;
     line-height: 96px;
@@ -154,13 +154,13 @@ export default {
     text-align: center;
     background-color: #fff;
   }
-  .u-actionsheet-item {
+  .lina-actionsheet-item {
     border-bottom: 1px solid #f6f6f6;
   }
-  .u-actionsheet-item-active {
+  .lina-actionsheet-item-active {
     color: #F0250F;
   }
-  .u-actionsheet-cancel {
+  .lina-actionsheet-cancel {
     margin-top: 10px;
     @include border-width-1px(1, 0, 0, 0, #f6f6f6);
   }